feat: added validator endpoint
