+++ src/worker.ts
@@ -1,28 +1,36 @@
-import { Value } from "@sinclair/typebox/value";
 import manifest from "../manifest.json";
 import { startStopTask } from "./plugin";
-import { Env, envConfigValidator, startStopSchema, StartStopSettings, startStopSettingsValidator } from "./types";
+import { Env } from "./types";
+import { validateAndDecodeSchemas } from "./utils/validator";
 
 export default {
   async fetch(request: Request, env: Env): Promise<Response> {
     try {
-      if (request.method === "GET") {
-        const url = new URL(request.url);
-        if (url.pathname === "/manifest.json") {
+      const url = new URL(request.url);
+      if (url.pathname === "/manifest") {
+        if (request.method === "GET") {
           return new Response(JSON.stringify(manifest), {
             headers: { "content-type": "application/json" },
           });
+        } else if (request.method === "POST") {
+          const webhookPayload = await request.json();
+
+          const result = validateAndDecodeSchemas(env, webhookPayload.settings);
+          if (result instanceof Response) {
+            return result;
+          }
+          return new Response(JSON.stringify({ message: "Schema is valid" }), { status: 200, headers: { "content-type": "application/json" } });
         }
       }
       if (request.method !== "POST") {
-        return new Response(JSON.stringify({ error: `Only POST requests are supported.` }), {
+        return new Response(JSON.stringify({ message: `Only POST requests are supported.` }), {
           status: 405,
           headers: { "content-type": "application/json", Allow: "POST" },
         });
       }
       const contentType = request.headers.get("content-type");
       if (contentType !== "application/json") {
-        return new Response(JSON.stringify({ error: `Error: ${contentType} is not a valid content type` }), {
+        return new Response(JSON.stringify({ message: `Error: ${contentType} is not a valid content type` }), {
           status: 400,
           headers: { "content-type": "application/json" },
         });
@@ -30,41 +38,16 @@ export default {
 
       const webhookPayload = await request.json();
 
-      const settings = Value.Default(startStopSchema, webhookPayload.settings) as StartStopSettings;
+      const result = validateAndDecodeSchemas(env, webhookPayload.settings);
 
-      if (!startStopSettingsValidator.test(settings)) {
-        const errorDetails: string[] = [];
-        for (const error of startStopSettingsValidator.errors(settings)) {
-          const errorMessage = `${error.path}: ${error.message}`;
-          console.error(errorMessage);
-          errorDetails.push(errorMessage);
-        }
-        return new Response(JSON.stringify({ error: `Bad Request: the settings are invalid. ${errorDetails.join("; ")}` }), {
-          status: 400,
-          headers: { "content-type": "application/json" },
-        });
+      if (result instanceof Response) {
+        return result;
       }
 
-      const decodedSettings = Value.Decode(startStopSchema, settings);
-
-      if (!envConfigValidator.test(env)) {
-        const errorDetails: string[] = [];
-        for (const error of envConfigValidator.errors(env)) {
-          const errorMessage = `${error.path}: ${error.message}`;
-          console.error(errorMessage);
-          errorDetails.push(errorMessage);
-        }
-        return new Response(JSON.stringify({ error: `Bad Request: the environment is invalid. ${errorDetails.join("; ")}` }), {
-          status: 400,
-          headers: { "content-type": "application/json" },
-        });
-      }
-
-      const decodedEnv = Value.Decode(envConfigValidator.schema, env);
-      webhookPayload.env = decodedEnv;
-      webhookPayload.settings = decodedSettings;
-      await startStopTask(webhookPayload, decodedEnv);
-      return new Response(JSON.stringify("OK"), { status: 200, headers: { "content-type": "application/json" } });
+      webhookPayload.env = result.decodedEnv;
+      webhookPayload.settings = result.decodedSettings;
+      await startStopTask(webhookPayload, result.decodedEnv);
+      return new Response(JSON.stringify({ message: "OK" }), { status: 200, headers: { "content-type": "application/json" } });
     } catch (error) {
       return handleUncaughtError(error);
     }
